[ET-VK] Enforce GPU buffer limit when partitioning
Pull Request resolved: https://github.com/pytorch/executorch/pull/6829

## Context

In Vulkan, there is a limit on the number of elements a GPU buffer can have. If a GPU buffer exceeds this limit, then the API will either produce an error or undefined behaviour will ensue.

## Changes

Along with `texture_limits`, introduce a configurable `buffer_limit` entry in the partitioner configuration.

Differential Revision: [D65899828](https://our.internmc.facebook.com/intern/diff/D65899828/)
ghstack-source-id: 253425728